24/03/25
Finished read hotels file and function to store hotel information in a hasmap.
Started on room class and reading rooms from a file.  ./resources/Rooms.txt
Need to figure out how to connect rooms to specific hotel if we are going to keep the roomsAvailable variable in the hotels class.